+++ backend/src/last-wills/lastwills.controller.ts
@@ -25,11 +25,11 @@ import {
   ApiUnauthorizedResponse,
 } from '@nestjs/swagger'
 import { LastWill, LastWillMetadata } from '../db/entities/lastwill.entity'
-import { LastWillsService } from '../db/services/lastwills.service'
 import { JwtGuard } from '../shared/guards/jwt.guard'
 import { RequestWithJWTPayload } from '../shared/interfaces/request-with-user.interface'
 import { CreateLastWillDto } from './dto/create-lastwill.dto'
 import { UpdateLastWillDto } from './dto/update-lastwill.dto'
+import { LastWillsService } from './lastwills.service'
 
 @UseGuards(JwtGuard)
 @ApiTags('lastwill')
